+++ src/main/java/roomescape/auth/dto/TokenRequest.java
@@ -1,23 +1,4 @@
 package roomescape.auth.dto;
 
-public class TokenRequest {
-
-    private String email;
-    private String password;
-
-    public TokenRequest() {
-    }
-
-    public TokenRequest(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
+public record TokenRequest(String email, String password) {
 }
