+++ src/main/java/roomescape/auth/exception/AuthorizationException.java
@@ -1,0 +1,4 @@
+package roomescape.auth.exception;
+
+public class AuthorizationException extends RuntimeException {
+}
