+++ src/main/java/roomescape/service/dto/MemberResponse.java
@@ -1,4 +1,4 @@
-package roomescape.auth.dto;
+package roomescape.service.dto;
 
 import roomescape.domain.Member;
